everything updated, adentu_real is float and adentu_int is int. Compiles with warnings, atom-pos-cuda.cu is giving problems at runtime.

+++ src/adentu-model.h
@@ -35,8 +35,8 @@ typedef struct _AdentuModel {
     double totalTime;
     double elapsedTime;
 
-    double gTemp;
-    double fTemp;
+    float gTemp;
+    float fTemp;
 
     vec3f vcmGrain;
     vec3f vcmFluid;
